move tests of IRVE_checker in the IRVE Analyser
